split lower_storage_key_limit tests
