+++ integration-tests/src/tests/client/features.rs
@@ -5,4 +5,5 @@ mod account_id_in_function_call_permission;
 #[cfg(feature = "protocol_feature_fix_contract_loading_cost")]
 mod fix_contract_loading_cost;
 mod fix_storage_usage;
+mod lower_storage_key_limit;
 mod restore_receipts_after_fix_apply_chunks;
