+++ integration-tests/src/tests/client/process_blocks.rs
@@ -103,7 +103,7 @@ pub fn create_nightshade_runtimes(genesis: &Genesis, n: usize) -> Vec<Arc<dyn Ru
 
 /// Produce `blocks_number` block in the given environment, starting from the given height.
 /// Returns the first unoccupied height in the chain after this operation.
-fn produce_blocks_from_height_with_protocol_version(
+pub(crate) fn produce_blocks_from_height_with_protocol_version(
     env: &mut TestEnv,
     blocks_number: u64,
     height: BlockHeight,
@@ -4518,145 +4518,5 @@ mod chunk_nodes_cache_test {
         assert_eq!(tx_node_counts[1], TrieNodesCount { db_reads: 12, mem_reads: 0 });
         assert_eq!(tx_node_counts[2], TrieNodesCount { db_reads: 8, mem_reads: 4 });
         assert_eq!(tx_node_counts[3], TrieNodesCount { db_reads: 8, mem_reads: 4 });
-    }
-}
-
-mod lower_storage_key_limit_test {
-    use super::*;
-
-    /// Check correctness of the protocol upgrade and ability to write 2 KB keys.
-    #[test]
-    fn protocol_upgrade() {
-        init_test_logger();
-
-        let old_protocol_version =
-            near_primitives::version::ProtocolFeature::LowerStorageKeyLimit.protocol_version() - 1;
-        let new_protocol_version = old_protocol_version + 1;
-        let new_storage_key_limit = 2usize.pow(11); // 2 KB
-        let args: Vec<u8> = vec![1u8; new_storage_key_limit + 1]
-            .into_iter()
-            .chain(near_primitives::test_utils::encode(&[10u64]).into_iter())
-            .collect();
-        let epoch_length: BlockHeight = 5;
-
-        // Prepare TestEnv with a contract at the old protocol version.
-        let mut env = {
-            let mut genesis =
-                Genesis::test(vec!["test0".parse().unwrap(), "test1".parse().unwrap()], 1);
-            genesis.config.epoch_length = epoch_length;
-            genesis.config.protocol_version = old_protocol_version;
-            let chain_genesis = ChainGenesis::new(&genesis);
-            let runtimes: Vec<Arc<dyn RuntimeAdapter>> =
-                vec![Arc::new(nearcore::NightshadeRuntime::test_with_runtime_config_store(
-                    Path::new("."),
-                    create_test_store(),
-                    &genesis,
-                    TrackedConfig::AllShards,
-                    RuntimeConfigStore::new(None),
-                )) as Arc<dyn RuntimeAdapter>];
-            let mut env = TestEnv::builder(chain_genesis).runtime_adapters(runtimes).build();
-
-            deploy_test_contract_with_protocol_version(
-                &mut env,
-                "test0".parse().unwrap(),
-                near_test_contracts::base_rs_contract(),
-                epoch_length.clone(),
-                1,
-                old_protocol_version,
-            );
-            env
-        };
-
-        let signer = InMemorySigner::from_seed("test0".parse().unwrap(), KeyType::ED25519, "test0");
-        let tx = Transaction {
-            signer_id: "test0".parse().unwrap(),
-            receiver_id: "test0".parse().unwrap(),
-            public_key: signer.public_key(),
-            actions: vec![Action::FunctionCall(FunctionCallAction {
-                method_name: "write_key_value".to_string(),
-                args,
-                gas: 10u64.pow(14),
-                deposit: 0,
-            })],
-
-            nonce: 0,
-            block_hash: CryptoHash::default(),
-        };
-
-        // Run transaction writing storage key exceeding the limit. Check that execution succeeds.
-        {
-            let tip = env.clients[0].chain.head().unwrap();
-            let signed_tx = Transaction {
-                nonce: tip.height + 1,
-                block_hash: tip.last_block_hash,
-                ..tx.clone()
-            }
-            .sign(&signer);
-            let tx_hash = signed_tx.get_hash().clone();
-            env.clients[0].process_tx(signed_tx, false, false);
-            produce_blocks_from_height_with_protocol_version(
-                &mut env,
-                epoch_length,
-                tip.height + 1,
-                old_protocol_version,
-            );
-            let final_result = env.clients[0].chain.get_final_transaction_result(&tx_hash).unwrap();
-            assert_matches!(final_result.status, FinalExecutionStatus::SuccessValue(_));
-        }
-
-        env.upgrade_protocol(new_protocol_version);
-
-        // Re-run the transaction, check that execution fails.
-        {
-            let tip = env.clients[0].chain.head().unwrap();
-            let signed_tx =
-                Transaction { nonce: tip.height + 1, block_hash: tip.last_block_hash, ..tx }
-                    .sign(&signer);
-            let tx_hash = signed_tx.get_hash().clone();
-            env.clients[0].process_tx(signed_tx, false, false);
-            for i in 0..epoch_length {
-                let block = env.clients[0].produce_block(tip.height + i + 1).unwrap().unwrap();
-                env.process_block(0, block.clone(), Provenance::PRODUCED);
-            }
-            let final_result = env.clients[0].chain.get_final_transaction_result(&tx_hash).unwrap();
-            assert_matches!(
-                final_result.status,
-                FinalExecutionStatus::Failure(TxExecutionError::ActionError(_))
-            );
-        }
-
-        // Run transaction where storage key exactly fits the new limit, check that execution succeeds.
-        {
-            let args: Vec<u8> = vec![1u8; new_storage_key_limit]
-                .into_iter()
-                .chain(near_primitives::test_utils::encode(&[20u64]).into_iter())
-                .collect();
-            let tx = Transaction {
-                signer_id: "test0".parse().unwrap(),
-                receiver_id: "test0".parse().unwrap(),
-                public_key: signer.public_key(),
-                actions: vec![Action::FunctionCall(FunctionCallAction {
-                    method_name: "write_key_value".to_string(),
-                    args,
-                    gas: 10u64.pow(14),
-                    deposit: 0,
-                })],
-
-                nonce: 0,
-                block_hash: CryptoHash::default(),
-            };
-            let tip = env.clients[0].chain.head().unwrap();
-            let signed_tx =
-                Transaction { nonce: tip.height + 1, block_hash: tip.last_block_hash, ..tx }
-                    .sign(&signer);
-            let tx_hash = signed_tx.get_hash().clone();
-            env.clients[0].process_tx(signed_tx, false, false);
-            for i in 0..epoch_length {
-                let block = env.clients[0].produce_block(tip.height + i + 1).unwrap().unwrap();
-                env.process_block(0, block.clone(), Provenance::PRODUCED);
-            }
-            let final_result = env.clients[0].chain.get_final_transaction_result(&tx_hash).unwrap();
-            assert_matches!(final_result.status, FinalExecutionStatus::SuccessValue(_));
-        }
     }
 }
